chore/#53: 공백 삭제 및 타입 수정

+++ app/src/main/java/com/wearerommies/roomie/domain/entity/TourEntity.kt
@@ -3,12 +3,12 @@ package com.wearerommies.roomie.domain.entity
 import java.util.Date
 
 data class TourEntity(
-    val name: String ,
+    val name: String,
     val birth: Date,
     val gender: String,
     val phoneNumber: String,
     val preferredDate: Date,
     val message: String,
-    val roomId: Int,
-    val houseId: Int
+    val roomId: Long,
+    val houseId: Long
 )
